requirements.txt
Django==2.2
django-filter==2.4.0
djangorestframework==3.12.2
drf-base64==2.0
django-anymail==7.1.0 
Pillow==7.2
psycopg2==2.8.6
tornado==5.1.1
redis==3.2.0
celery==4.2.1
flower==0.9.2